requirements.txt
asgiref==3.8.1
Django==4.2.2
django-environ==0.12.0
django-hosts==6.0
django-storages==1.14.5
gunicorn==23.0.0
packaging==24.2
pillow==11.1.0
psycopg2==2.9.10
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
boto3